pyproject.toml
[project]
name = "rand-org-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bigtree[polars]>=0.25.2",
    "factory-boy>=3.3.3",
    "faker>=37.0.2",
]

[dependency-groups]
dev = [
    "icecream>=2.1.4",
]
